+++ joulescope/command/runner.py
@@ -19,10 +19,10 @@
 import sys
 import argparse
 import logging
-from joulescope.command import capture, capture_usb, program, recording
+from joulescope.command import capture, capture_usb, gpo_demo, program, recording
 
 
-commands = [capture, capture_usb, program, recording]
+commands = [capture, capture_usb, gpo_demo, program, recording]
 """This list of available command modules.  Each module must contain a 
 parser_config(subparser) function.  The function must return the callable(args)
 that will be executed for the command."""