+++ joulescope/usb/device_thread.py
@@ -25,7 +25,7 @@ from joulescope.usb.api import DeviceDriverApi
 log = logging.getLogger(__name__)
 log.setLevel(level=logging.INFO)
 TIMEOUT = 3.0
-TIMEOUT_OPEN = 6.0
+TIMEOUT_OPEN = 10.0
 
 
 def default_callback(*args, **kwargs):