+++ CHANGELOG.md
@@ -8,7 +8,8 @@ This file contains the list of changes made to pyjoulescope.
 
 2019 Jun ?? [in development]
 
-*   Added GPO alias values [0, 1]
+*   Added GPO alias values [0, 1].
+*   Added GPO demo exampe.