+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This file contains the list of changes made to pyjoulescope.
 *   Improved Device.bootloader error handling.
 *   Removed libusb IOError if platform does not support hotplug.
 *   Added no exception guarantee to scan() and bootloaders_run_application().
+*   Fixed Cython compiler warnings.
 
 
 ## 0.4.2